fixup! Create My riddle detail
